+++ PokeDex/PokeDex/Model Controller/PokemonController.swift
@@ -1,0 +1,9 @@
+//
+//  PokemonController.swift
+//  PokeDex
+//
+//  Created by Nichole Davidson on 4/10/20.
+//  Copyright © 2020 Nichole Davidson. All rights reserved.
+//
+
+import Foundation
