+++ PokeDex/PokeDex/Resources/Info.plist
